pyproject.toml
[tool.poetry]
name = "drf-api-authentication"
version = "0.1.0"
description = ""
authors = ["Moha-AlHanbali <moha.alhanbali@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
whitenoise = "^5.3.0"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.10.0"
gunicorn = "^20.1.0"
postgres = "^4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
